need longer ints for length calculation
git-svn-id: https://vgmstream.svn.sourceforge.net/svnroot/vgmstream@835 51a99a44-fe44-0410-b1ba-c3e57ba2b86b

+++ src/coding/mpeg_decoder.c
@@ -215,7 +215,7 @@ void decode_mpeg(VGMSTREAMCHANNEL *stream,
 }
 
 long mpeg_bytes_to_samples(long bytes, const struct mpg123_frameinfo *mi) {
-    return bytes * mi->rate * 8 / (mi->bitrate * 1000);
+    return (int64_t)bytes * mi->rate * 8 / (mi->bitrate * 1000);
 }
 
 #endif
